% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adv_update_prm.r
\name{adv_update_prm}
\alias{adv_update_prm}
\title{adv_update_prm}
\usage{
adv_update_prm(
  prm_file,
  parameter_names,
  new_values,
  indices,
  custom_details = NULL
)
}
\arguments{
\item{prm_file}{the bgm file that contains the parameter you want to change. Write as "bgm_file.prm".}

\item{parameter_name}{the parameter that you want to change. Write as "parameter_name".}

\item{new_value}{the new value for the parameter.}

\item{index}{the index number of the parameter value to change when dealing with multiple values per single parameter (e.g, pPrey, migration matrices).}
}
\value{
An updated bgm file with a new value for the parameter of interest
}
\description{
Advanced function for changing parameter value in a bgm file for the Atlantis model.
The function does not overwrite the old file, but creates a new one with details of the change in its name.
}
\examples{
Save new_value and indices as vectors with one or more values inside them.
E.g., for pPreyWHH, set new value <- 0.06 for prey index 28 (KR), and new value <- 0.04 for prey index 1 (FM):
new_values <- c(0.06, 0.04)
indices <- c(28, 1)
adv_update_prm(biology.prm, "pPreyWHH", new_values, indices, custom_details = "_ZGdiet_change")
}
