% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metric.r
\name{get_metric}
\alias{get_metric}
\title{get_metric function
Extract data for RelBiomass and input parameters from Atlantis output .txt and xml files.}
\usage{
get_metric(base_path)
}
\arguments{
\item{base_path}{The path to the folder containing the output folders.}
}
\value{
A list of datasets for relative biomass (final/initial biomass values) and parameter values from biology input xml file.
}
\description{
get_metric function
Extract data for RelBiomass and input parameters from Atlantis output .txt and xml files.
}
\examples{
get_metric("your/output/files/path")
}
