% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massplot.r
\name{mass_plot}
\alias{mass_plot}
\title{mass_plot function
Plot multiple instances of Atlantis output files.}
\usage{
mass_plot(base_path)
}
\arguments{
\item{base_path}{The path to the folder containing the output folders.}
}
\value{
A facet plot with one panel per functional group, overlying lines for different model runs.
}
\description{
mass_plot function
Plot multiple instances of Atlantis output files.
}
\examples{
mass_plot("your/output/files/path")
}
