% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rel_mass_plot.r
\name{rel_mass_plot}
\alias{rel_mass_plot}
\title{rel_mass_plot function
Plot multiple instances of Atlantis output files, using relative values.}
\usage{
rel_mass_plot(base_path)
}
\arguments{
\item{base_path}{The path to the folder containing the output folders.}
}
\value{
A facet plot with one panel per functional group, overlying lines for different model runs.
}
\description{
rel_mass_plot function
Plot multiple instances of Atlantis output files, using relative values.
}
\examples{
rel_mass_plot("your/output/files/path")
}
