% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_docker_instances.R
\name{run_docker_instances}
\alias{run_docker_instances}
\title{run_docker_instances}
\usage{
run_docker_instances(local_path_to_prm_files, docker_image)
}
\arguments{
\item{local_path_to_prm_files}{Local path where Docker should look for the .prm files, including the modified ones, to mount them into the container. The path needs to be in quotes ("my/file/path").}

\item{docker_image}{The name of the Docker image that you are using to run the model (e.g., atlantis_docker).}
}
\value{
Runs a new Docker container for each modified .prm file found in the specified directory.
}
\description{
run_docker_instances
}
\examples{
run_docker_instances(local_path_to_prm_files = "my/file/path", docker_image = 'atlantis_docker')
}
