% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whole_shebang.r
\name{whole_shebang}
\alias{whole_shebang}
\title{whole_shebang}
\usage{
whole_shebang(prm_file, base_path)
}
\arguments{
\item{base_path}{The path to the folder containing the output folders.}
}
\value{
Optimisation process for Atlantis model.
}
\description{
whole_shebang
}
